+++ mayan/apps/smart_settings/apps.py
@@ -1,14 +1,16 @@
 from django.utils.translation import ugettext_lazy as _
 
 from mayan.apps.common.apps import MayanAppConfig
-from mayan.apps.common.menus import menu_secondary, menu_setup, menu_object
+from mayan.apps.common.menus import (
+    menu_list_facet, menu_object, menu_return, menu_setup
+)
 from mayan.apps.navigation.classes import SourceColumn
 from mayan.apps.views.html_widgets import TwoStateWidget
 
 from .classes import SettingNamespace, Setting
 from .links import (
-    link_setting_namespace_detail, link_setting_namespace_list, link_namespace_root_list,
-    link_setting_edit
+    link_setting_namespace_detail, link_setting_namespace_list,
+    link_namespace_root_list, link_setting_edit
 )
 from .widgets import setting_widget
 
@@ -43,15 +45,15 @@ class SmartSettingsApp(MayanAppConfig):
             widget=TwoStateWidget
         )
 
-        menu_object.bind_links(
+        menu_list_facet.bind_links(
             links=(link_setting_namespace_detail,), sources=(SettingNamespace,)
         )
         menu_object.bind_links(
             links=(link_setting_edit,), sources=(Setting,)
         )
-        menu_secondary.bind_links(
+        menu_return.bind_links(
             links=(link_namespace_root_list,), sources=(
-                SettingNamespace, Setting, 'settings:setting_namespace_list',
+                SettingNamespace, Setting
             )
         )
         menu_setup.bind_links(links=(link_setting_namespace_list,))