+++ HISTORY.rst
@@ -118,6 +118,7 @@
 - Add .msg file metadata drivers.
 - Convert the metadata value and the file metadata value fields from
   character fields with a maximum length of 255 characters to text fields.
+- Improve the settings apps navigation.
 
 4.3.1 (2022-08-21)
 ==================